Add files via upload
Code to assist with flagging suspect data. Uses a generic "mtcars" example but can be shaped to whatever data and logic you would like to flag on. The real use is learning how to write a new flag column, and add to this if flagging logic overlaps. These flags can then be checked 1 by 1 in excel or whatever medium when looking back through data sheets to try to find entry errors.